Add test file for item count
