+++ jest.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  testEnvironment: 'jsdom',
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+  },
+};
