+++ modules/setEvents.js
